Add the PUT, POST, and DELETE routes
In the package.json and package-lock.json files, install body-parser. In the server.js file, import and use body-parser. In the routes/contactsRoutes.js file, add the post, put, and delete routes. In the controllers/contactsController.js file, create the createContact, updateContact, and deleteContact functions. And in the requests.rest file, test the routes.